Updated code formatting

+++ src/system/application/admin/controllers/CategoryController.php
@@ -1,13 +1,21 @@
 <?php
 
-class Admin_CategoryController extends SimpleCart_AdminController {
+/**
+ * Catagory management
+ *
+ * @package Simplecart
+ * @subpackage AdminController
+ * @author spekkionu
+ */
+class Admin_CategoryController extends SimpleCart_AdminController
+{
 
-  public function init(){
+  public function init() {
     parent::init();
-    if(!$this->isAllowed('admin', 'general')){
-      if(Zend_Auth::getInstance()->hasIdentity()){
+    if (!$this->isAllowed('admin', 'general')) {
+      if (Zend_Auth::getInstance()->hasIdentity()) {
         return $this->_forward('denied', 'access', 'admin');
-      }else{
+      } else {
         // Save current url for later
         $session = new Zend_Session_Namespace();
         $session->login_destination = $this->view->url();
@@ -16,68 +24,79 @@ class Admin_CategoryController extends SimpleCart_AdminController {
     }
   }
 
-  public function indexAction(){
+  public function indexAction() {
+    if (!$this->isAllowed('admin:category', 'view')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
     $categories = Doctrine::getTable('Category')->getCategoryTree();
     $this->view->categories = $categories;
-
   }
 
-  public function detailsAction(){
-    if($this->getRequest()->isXmlHttpRequest()){
+  public function detailsAction() {
+    if ($this->getRequest()->isXmlHttpRequest()) {
       // Disable Layout for AJAX Requests
       $this->_helper->layout->disableLayout();
     }
+    if (!$this->isAllowed('admin:category', 'view')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
     $id = intval($this->getRequest()->getParam('id'));
-    if(!$id){
+    if (!$id) {
       return $this->_forward('not-found', 'error', 'default');
     }
     $category = Doctrine::getTable('Category')->getCategory($id);
-    if(!$category){
+    if (!$category) {
       return $this->_forward('not-found', 'error', 'default');
     }
     $this->view->category = $category;
   }
 
-  public function moveAction(){
-    if(!$this->getRequest()->isXmlHttpRequest()){
+  public function moveAction() {
+    if (!$this->getRequest()->isXmlHttpRequest()) {
       return $this->_forward('not-found', 'error', 'default');
     }
-    if(!$this->getRequest()->isPost()){
-      return $this->sendJson(array('success'=>false,'message'=>'This must be a POST request.'));
+    if (!$this->isAllowed('admin:category', 'edit')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
+    if (!$this->getRequest()->isPost()) {
+      return $this->sendJson(array('success' => false, 'message' => 'This must be a POST request.'));
     }
     $id = $this->getRequest()->getPost('category');
     $position = $this->getRequest()->getPost('position');
     $ref = $this->getRequest()->getPost('ref');
-    try{
+    try {
       Doctrine::getTable('Category')->moveCategory($id, $ref, $position);
-      return $this->sendJson(array('success'=>true,'message'=>'Successfully moved category.'));
-    }catch(Exception $e){
-      return $this->sendJson(array('success'=>false,'message'=>'Failed to move category.'));
+      return $this->sendJson(array('success' => true, 'message' => 'Successfully moved category.'));
+    } catch (Exception $e) {
+      return $this->sendJson(array('success' => false, 'message' => 'Failed to move category.'));
     }
   }
 
-  public function addAction(){
+  public function addAction() {
+    if (!$this->isAllowed('admin:category', 'add')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
     $id = $this->getRequest()->getParam('parent');
     $parent = Doctrine::getTable('Category')->getCategory($id);
-    if(!$parent){
+    if (!$parent) {
       $id = NULL;
     }
     $form = new Form_Category();
     $form->addDbValidators();
-    if($id){
+    if ($id) {
       $form->getElement('parent')->setValue($id);
     }
-    if($this->getRequest()->isPost()){
-      if($form->isValid($this->getRequest()->getPost())){
-        try{
+    if ($this->getRequest()->isPost()) {
+      if ($form->isValid($this->getRequest()->getPost())) {
+        try {
           $values = $form->getValues();
           Doctrine::getTable('Category')->addCategory($values);
           $this->addMessage("Category {$values['name']} added.", 'success');
           return $this->routeRedirect("admin_category");
-        }catch(Exception $e){
+        } catch (Exception $e) {
           $this->logError("Error adding new category. - {$e->getMessage()}");
           $this->addMessage("Error adding category", 'error');
-          return $this->routeRedirect('admin_category_add', array('parent'=>$values['parent']));
+          return $this->routeRedirect('admin_category_add', array('parent' => $values['parent']));
         }
       }
       // Refresh parent category
@@ -93,30 +112,33 @@ class Admin_CategoryController extends SimpleCart_AdminController {
     // Activate Menu Items
     $container = Zend_Registry::get('Zend_Navigation');
     $found = $container->findOneByRoute('admin_category_add');
-    $found->set('params', array('parent'=>$id));
+    $found->set('params', array('parent' => $id));
   }
 
-  public function editAction(){
+  public function editAction() {
+    if (!$this->isAllowed('admin:category', 'edit')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
     $id = $this->getRequest()->getParam('id');
     $category = Doctrine::getTable('Category')->find($id);
-    if(!$id){
+    if (!$id) {
       return $this->_forward('not-found', 'error', 'default');
     }
     $form = new Form_Category();
     $form->removeElement('parent');
     $form->addDbValidators($category->route_id);
     $form->populate($category->toArray());
-    if($this->getRequest()->isPost()){
-      if($form->isValid($this->getRequest()->getPost())){
-        try{
+    if ($this->getRequest()->isPost()) {
+      if ($form->isValid($this->getRequest()->getPost())) {
+        try {
           $values = $form->getValues();
           Doctrine::getTable('Category')->updateCategory($category, $values);
           $this->addMessage("Category {$values['name']} updated.", 'success');
           return $this->routeRedirect("admin_category");
-        }catch(Exception $e){
+        } catch (Exception $e) {
           $this->logError("Error updating category. - {$e->getMessage()}");
           $this->addMessage("Error updating category", 'error');
-          return $this->routeRedirect('admin_category_edit', array('id'=>$id));
+          return $this->routeRedirect('admin_category_edit', array('id' => $id));
         }
       }
     }
@@ -127,28 +149,31 @@ class Admin_CategoryController extends SimpleCart_AdminController {
     $container = Zend_Registry::get('Zend_Navigation');
     $found = $container->findOneByRoute('admin_category_edit');
     $found->setVisible(true);
-    $found->set('params', array('id'=>$id));
+    $found->set('params', array('id' => $id));
   }
 
-  public function deleteAction(){
+  public function deleteAction() {
+    if (!$this->isAllowed('admin:category', 'delete')) {
+      return $this->_forward('denied', 'access', 'admin');
+    }
     $id = $this->getRequest()->getParam('id');
     $category = Doctrine::getTable('Category')->find($id);
-    if(!$id){
+    if (!$id) {
       return $this->_forward('not-found', 'error', 'default');
     }
     $can_delete = !$category->getNode()->hasChildren();
-    if($can_delete){
+    if ($can_delete) {
       $form = new Form_Delete();
-      if($this->getRequest()->isPost()){
-        if($form->isValid($this->getRequest()->getPost())){
-          try{
+      if ($this->getRequest()->isPost()) {
+        if ($form->isValid($this->getRequest()->getPost())) {
+          try {
             Doctrine::getTable('Category')->deleteCategory($category);
             $this->addMessage("Category {$category['name']} deleted.", 'success');
             return $this->routeRedirect("admin_category");
-          }catch(Exception $e){
+          } catch (Exception $e) {
             $this->logError("Error deleting category. - {$e->getMessage()}");
             $this->addMessage("Error deleting category", 'error');
-            return $this->routeRedirect('admin_category_delete', array('id'=>$id));
+            return $this->routeRedirect('admin_category_delete', array('id' => $id));
           }
         }
       }
@@ -160,6 +185,7 @@ class Admin_CategoryController extends SimpleCart_AdminController {
     $container = Zend_Registry::get('Zend_Navigation');
     $found = $container->findOneByRoute('admin_category_delete');
     $found->setVisible(true);
-    $found->set('params', array('id'=>$id));
+    $found->set('params', array('id' => $id));
   }
+
 }
